add api module for occurrences
